% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exim.R
\name{exim_by_year}
\alias{exim_by_year}
\title{Summarize EXIM authorizations by fiscal year}
\usage{
exim_by_year()
}
\value{
tibble: fiscal_year, n_deals, total_approved, total_disbursed
}
\description{
Summarize EXIM authorizations by fiscal year
}
